transaction_service/requirements.txt
asgiref==3.8.1
cffi==1.17.1
confluent-kafka==2.3.0
cryptography==44.0.2
Django==5.1.7
django-environ==0.12.0
djangorestframework==3.15.2
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
setuptools==76.0.0
sqlparse==0.5.3
